# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cavendo Engine API
  description: |
    REST API for Cavendo Engine - an AI agent task management platform.

    ## Authentication

    The API supports three authentication methods:

    ### Session Authentication (Cookie-based)
    Used by the web UI. Login via `/api/auth/login` to receive a session cookie.

    ### Agent Key Authentication (X-Agent-Key header)
    Used by AI agents. Generate keys via the admin UI or `/api/agents/{id}/keys`.
    Keys start with `cav_ak_` prefix.

    ### User Key Authentication (X-Agent-Key header)
    Personal API keys for users (e.g., for MCP integration). Generate via `/api/users/me/keys`.
    Keys start with `cav_uk_` prefix.

    ## Key Concepts

    - **Agents**: AI agents that execute tasks (Claude, GPT, custom implementations)
    - **Tasks**: Work items assigned to agents
    - **Deliverables**: Work output submitted by agents for review
    - **Projects**: Organize tasks and knowledge
    - **Knowledge**: Project-scoped reference documents for agents
    - **Webhooks**: Event notifications for agents

  version: 0.1.0
  contact:
    name: Cavendo Engine
    url: https://github.com/Cavendo/Engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://your-server.com
    description: Production
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie obtained from login endpoint
    agentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key (cav_ak_...) or User API key (cav_uk_...)
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            errors:
              type: object
              additionalProperties:
                type: string
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        taskCounts:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            in_progress:
              type: integer
            pending:
              type: integer
    ProjectWithTasks:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            tasks:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  status:
                    type: string
                  priority:
                    type: integer
                  assigned_agent_id:
                    type: integer
                    nullable: true
            knowledgeCount:
              type: integer
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
    UpdateProjectRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        status:
          type: string
          enum: [active, archived, completed]

    # Knowledge schemas
    Knowledge:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
        content_type:
          type: string
          enum: [markdown, html, json, text]
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        project_name:
          type: string
          description: Populated when querying with joins
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Session required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List all projects
      description: Returns projects with task counts. Accessible by users and agents.
      security:
        - sessionAuth: []
        - agentKey: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, completed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create a new project
      description: Creates a new project. Requires admin role.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      description: Returns project details including tasks. Accessible by users and agents.
      security:
        - sessionAuth: []
        - agentKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ProjectWithTasks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update project
      description: Updates project properties. Requires admin role.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Projects]
      summary: Delete project
      description: Deletes a project. Cannot delete if tasks are associated. Requires admin role.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                        example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Cannot delete project with associated tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/projects/{id}/knowledge:
    get:
      tags: [Projects]
      summary: Get project knowledge base
      description: Returns knowledge entries for a project. Accessible by users and agents.
      security:
        - sessionAuth: []
        - agentKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Project knowledge
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      project:
                        $ref: '#/components/schemas/Project'
                      knowledge:
                        type: array
                        items:
                          $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Knowledge endpoints
...